trying out functions
Still not able to summarize all the expenses. I believe the problem is that the array contains both strings and numbers. Trying out different solutions for this.

Created a graveyard to store code that I've been trying out so to remember what I have already tried and not since I feel like I'm going in circles.