problem solved
with help from teacher assistent I managed to make a function that summarized all the numbers in the expences-array! Turns out I had created a similar function before but put a keyword in the wrong position.

Also got help to make the "landing page" of menu come after each input. Next step is to make sure the break function works.